requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
web3==6.15.1
bitcoinlib==0.6.13
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
redis==5.0.1
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.9
aiohttp==3.9.3
websockets==12.0
pytest==7.4.4
httpx==0.27.0
neo4j==5.16.0
influxdb-client==1.39.0
